+++ example/src/commands/hello/index.ts
@@ -1,8 +1,9 @@
 import { Flags } from '@oclif/core'
 
+import type { InferFlags } from '@cenk1cenk2/oclif-common'
 import { Command } from '@cenk1cenk2/oclif-common'
 
-export default class Hello extends Command {
+export default class Hello extends Command<any, InferFlags<typeof Hello>> {
   static description = 'Say hello'
 
   static examples = [ '$ oex hello friend --from oclif' ]
@@ -24,8 +25,6 @@ export default class Hello extends Command {
   ]
 
   public async run (): Promise<void> {
-    const { args, flags } = await this.parse(Hello)
-
-    this.logger.info(`hello ${args.person} from ${flags.from}! (./src/commands/hello/index.ts)`)
+    this.logger.info(`hello ${this.args.person} from ${this.flags.from}! (./src/commands/hello/index.ts)`)
   }
 }