+++ packages/common/src/index.ts
@@ -1,6 +1,12 @@
+import { CliUx as BaseCliUx } from '@oclif/core'
+
 export * from './commands'
 export * from './constants'
 export * from './hooks'
 export * from './interfaces'
 export * from './lib'
 export * from './utils'
+
+export default {
+  ux: BaseCliUx.ux
+}