+++ packages/common/src/interfaces/index.ts
@@ -1,2 +1,3 @@
 export * from './class.interface'
 export * from './config-command.interface'
+export * from './oclif.interface'