+++ packages/common/src/lib/parser/index.ts
@@ -1,2 +1,4 @@
 export * from './parser.interface'
+export * from './parser.service'
 export * from './yaml-parser.service'
+export * from './json-parser.service'